Store the accepted counts with an integer dtype
The in-memory backend allocated the per-walker acceptance counter with
the chain dtype and the HDF backend with float64; both now use int64,
which matches what the counter holds. acceptance_fraction still
divides into a float array.

+++ src/emcee/backends/backend.py
@@ -49,7 +49,9 @@ class Backend:
         self.nwalkers = int(nwalkers)
         self.ndim = int(ndim)
         self.iteration = 0
-        self.accepted = np.zeros(self.nwalkers, dtype=self.dtype)
+        # A per-walker count of accepted proposals, not chain data, so
+        # it gets an integer dtype independent of ``self.dtype``
+        self.accepted = np.zeros(self.nwalkers, dtype=np.int64)
         self.chain = np.empty((0, self.nwalkers, self.ndim), dtype=self.dtype)
         self.log_prob = np.empty((0, self.nwalkers), dtype=self.dtype)
         self.blobs = None
